consolidate response struct within device
