+++ device.go
@@ -79,3 +79,12 @@ type State struct {
 	Duration   float64 `json:"duration"`
 	Infrared   float64 `json:"infrared"`
 }
+
+// Response is a generic slice of results from LIFX API
+type Response struct {
+	Results []struct {
+		ID     string `json:"id"`
+		Status string `json:"status"`
+		Label  string `json:"label"`
+	} `json:"results"`
+}
